Separate independent task for extraction.
