+++ apps/etl/extract.py
@@ -12,6 +12,8 @@ class Extraction:
         resp_type = response.headers.get("Content-Type", "")
         file_extension = "txt"
         resp_data = response
+        resp_status = ExtractionData.Status.IN_PROGRESS
+        source_validation_status = ExtractionData.ValidationStatus.NO_VALIDATION
 
         if response.status_code == 200:
             if "application/json" in resp_type:
@@ -47,13 +49,16 @@ class Extraction:
                 "source_validation_status": ExtractionData.ValidationStatus.NO_VALIDATION,
                 "content_validation": "",
             }
+        elif response.status_code == 204:
+            resp_status = ExtractionData.Status.SUCCESS
+            source_validation_status = ExtractionData.ValidationStatus.NO_DATA
         return {
             "source": source,
             "url": self.url,
             "attempt_no": attempt_no,
             "resp_code": response.status_code,
-            "status": ExtractionData.Status.FAILED,
-            "source_validation_status": ExtractionData.ValidationStatus.NO_VALIDATION,
+            "status": resp_status,
+            "source_validation_status": source_validation_status,
             "resp_data": None,
             "resp_data_type": "",
             "file_extension": None,
